+++ src/Phan/Issue.php
@@ -578,6 +578,7 @@ class Issue
     public const CompatibleTrailingCommaParameterList = 'PhanCompatibleTrailingCommaParameterList';
     public const CompatibleAttributeGroupOnSameLine      = 'PhanCompatibleAttributeGroupOnSameLine';
     public const CompatibleAttributeGroupOnMultipleLines = 'PhanCompatibleAttributeGroupOnMultipleLines';
+    public const CompatibleConstructorPropertyPromotion = 'PhanCompatibleConstructorPropertyPromotion';
 
     // Issue::CATEGORY_GENERIC
     public const TemplateTypeConstant       = 'PhanTemplateTypeConstant';
@@ -4315,6 +4316,7 @@ class Issue
                 self::REMEDIATION_B,
                 8005
             ),
+            // FIXME: It's redundant to include the first FILE:LINE of the declaration in the full issue message
             new Issue(
                 self::RedefineClassAlias,
                 self::CATEGORY_REDEFINE,
@@ -4963,12 +4965,20 @@ class Issue
                 3037
             ),
             new Issue(
+                self::CompatibleConstructorPropertyPromotion,
+                self::CATEGORY_COMPATIBLE,
+                self::SEVERITY_NORMAL,
+                "Cannot use constructor property promotion before php 8.0 for {PARAMETER} of {METHOD}",
+                self::REMEDIATION_B,
+                3038
+            ),
+            new Issue(
                 self::CompatibleAttributeGroupOnSameLine,
                 self::CATEGORY_COMPATIBLE,
                 self::SEVERITY_CRITICAL,
                 "Declaring attributes on the same line as a declaration is treated like a line comment before php 8.0 for attribute group {CODE} of {CODE}",
                 self::REMEDIATION_B,
-                3038
+                3039
             ),
             new Issue(
                 self::CompatibleAttributeGroupOnMultipleLines,
@@ -4976,7 +4986,7 @@ class Issue
                 self::SEVERITY_CRITICAL,
                 "Declaring attributes across multiple lines may be treated like a mix of a line comment and php tokens before php 8.0 for attribute group {CODE} of {CODE} ending around line {LINE}. Note that php-ast does not provide the actual ending line numbers and this issue may be unreliable",
                 self::REMEDIATION_B,
-                3039
+                3040
             ),
 
             // Issue::CATEGORY_GENERIC