+++ tests/php80_files/src/032_variadic_promoted_property.php
@@ -1,0 +1,10 @@
+<?php
+class VariadicPromotedProperty {
+    public function __CONSTRUCT(public int $value, public ...$rest) {
+    }
+
+    public function other(public int $foo) {
+    }
+}
+$c = fn(public int $x) => 123;
+function fn32(public $var) {}