+++ tests/php80_files/src/034_promoted_property_infer.php
@@ -1,0 +1,16 @@
+<?php
+class HasUntypedProperty {
+    public $x;
+    public function __construct($x) {
+        $this->x = $x;
+    }
+}
+class HasUntypedPropertyPromotion {
+    public function __construct(public $x) {
+        echo strlen($this->x);
+    }
+}
+$c = new HasUntypedProperty(new stdClass());
+echo strlen($c->x);
+$c = new HasUntypedPropertyPromotion([]);
+echo strlen($c->x);