+++ NEWS.md
@@ -1,5 +1,10 @@
 Phan NEWS
 
+Nov 26 2020, Phan 4.0.0-alpha5 (dev)
+------------------------------
+
+Merge changes from Phan 3.2.6.
+
 Nov 26 2020, Phan 4.0.0-alpha4
 ------------------------------
 
@@ -34,6 +39,17 @@ Backwards incompatible changes:
 
 Miscellaneous:
 + Make various classes from Phan implement `Stringable`.
+
+Nov 26 2020, Phan 3.2.6 (dev)
+-----------------------
+
+New features (Analysis):
++ Update many more real parameter names to match php 8.0's parameter names for php's own internal methods. (#4263)
++ Infer that an instance property exists for PHP 8.0 constructor property promotion. (#3938)
++ Infer types of properties from arguments passed into constructor for PHP 8.0 constructor property promotion. (#3938)
++ Emit `PhanInvalidNode` and `PhanRedefineProperty` when misusing syntax for constructor property promotion. (#3938)
++ Emit `PhanCompatibleConstructorPropertyPromotion` when the project's `minimum_target_php_version` is older than `8.0` (#3938)
++ Emit `PhanSuspiciousMagicConstant` when using `__FUNCTION__` inside of a closure. (#4222)
 
 Nov 26 2020, Phan 3.2.5
 -----------------------